5/23/2024 fcm added but milestone start date and end date are issue

+++ src/main/java/com/task_pay/task_pay/models/dtos/TaskDto.java
@@ -21,7 +21,7 @@ public class TaskDto {
     private String taskName;
 
     @NotNull(message = "required taskPrice !")
-    private Integer taskPrice;
+    private double taskPrice;
 
     @NotBlank(message = "required taskAbout !")
     private String taskAbout;
